Mark episode as retreived on https://www.betaseries.com/ after download
To be able to mark an episaode as retreived tget need to get a user token on https://www.betaseries.com/
This is done automaticly if the user and password is of betaseries is present in the config file

config file has been refactored and config command modify. Now we expect the user to create manually the config file.